Limit that only changesets lower than HEAD can be deleted
- Add new GetHeadVersion QR which fetches the max version from the space table
- Change DeleteChangesets QR to throw an ILLEGAL_ARGUMENT error if the user tries to specify a value larger than HEAD (uses new GetHeadVersion QR)
- Add test deleteChangesetsLargerThanHeadValue() which tests that the above error is thrown
- Also refactor all occurrences of "version" to be long rather than int

+++ xyz-hub-test/src/test/java/com/here/xyz/hub/rest/ReadHistoryApiIT.java
@@ -650,7 +650,7 @@ public class ReadHistoryApiIT extends TestSpaceWithFeature {
 
   private void checkCompactChangesetVersionIntegrity(List<Feature> fList) {
     for (Feature f : fList) {
-      int v = f.getProperties().getXyzNamespace().getVersion();
+      long v = f.getProperties().getXyzNamespace().getVersion();
       if(v < 11) {
         assertEquals(true, f.getProperties().get("free"));
       }if(v == 11 || v == 12){
